requirements.txt
antlr4-python3-runtime==4.10
antlr4-verilog==4.0.0
bleach==5.0.1
build==0.8.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
commonmark==0.9.1
cryptography==37.0.4
cycler==0.11.0
docutils==0.19
fonttools==4.34.4
graphviz==0.20.1
idna==3.3
importlib-metadata==4.12.0
jeepney==0.8.0
keyring==23.8.2
kiwisolver==1.4.4
matplotlib==3.5.2
networkx==2.8.5
numpy==1.23.1
packaging==21.3
pandas==1.4.3
pep517==0.13.0
Pillow==9.2.0
pkginfo==1.8.3
pycparser==2.21
pydot==1.4.2
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
readme-renderer==36.0
requests==2.28.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.5.1
scipy==1.9.0
seaborn==0.11.2
SecretStorage==3.3.2
six==1.16.0
tomli==2.0.1
tqdm==4.64.0
twine==4.0.1
urllib3==1.26.11
webencodings==0.5.1
zipp==3.8.1
